==> research/04_trading_strategy.ipynb <==
import numpy as np
import pandas as pd

df_ml = pd.read_csv("../data/data_ml.csv")

separation_date = "2013-12-31"
df_train = df_ml.query("date < @separation_date").copy()
df_test = df_ml.query("@separation_date <= date & date < '2018-12-31'").copy()

from sklearn.linear_model import LogisticRegression

columns_to_drop = [
    'stock_id', 'date', # non-feature identifiers
    'R1M_Usd', 'R3M_Usd', 'R6M_Usd', 'R12M_Usd', # numerical labels
    'R1M_Usd_C', 'R12M_Usd_C' # categorical labels
]

features = list(df_ml.drop(columns=columns_to_drop).columns)

X_train = df_train[features]
y_train = df_train[['R1M_Usd_C']]

X_test = df_test[features]
y_test = df_test[['R1M_Usd_C']]

from sklearn.linear_model import LogisticRegression

model = LogisticRegression()

from sklearn.model_selection import cross_val_score

cross_val_score(model, X_train, y_train.values.ravel())

model.fit(X_train, y_train.values.ravel())

model.predict(X_test).mean()

np.mean(model.predict_proba(X_test)[:,0] > 0.5)

df_probabilities = pd.DataFrame(model.predict_proba(X_test))
df_probabilities

df_test['prediction_hard'] = model.predict(X_test)
df_test

df_test['predict_prob_0'] = model.predict_proba(X_test)[:,0]
df_test['predict_prob_1'] = model.predict_proba(X_test)[:,1]
df_test

df_monthly_returns = df_test.query("prediction_hard == 1").groupby("date")[["R1M_Usd"]].mean().reset_index()
df_monthly_returns['growth_factor'] = 1 + df_monthly_returns['R1M_Usd']
df_monthly_returns['equity_curve'] = df_monthly_returns['growth_factor'].cumprod()
df_monthly_returns

(df_monthly_returns["R1M_Usd"].mean() / df_monthly_returns["R1M_Usd"].std()) * np.sqrt(12)

(1.585734 ** (1/5)) - 1
