==> research/06_random_forest_hyperparameter_tuning.ipynb <==
import numpy as np
import pandas as pd

df_ml = pd.read_csv("../data/data_ml.csv")

separation_date = "2013-12-31"
df_train = df_ml.query("date < @separation_date").copy()
df_test = df_ml.query("@separation_date <= date & date < '2018-12-31'").copy()

columns_to_drop = [
    "stock_id", "date", # non-feature identifiers
    "R1M_Usd", "R3M_Usd", "R6M_Usd", "R12M_Usd", # numerical labels
    "R1M_Usd_C", "R12M_Usd_C" # categorical labels
]

features = list(df_ml.drop(columns=columns_to_drop).columns)

X_train = df_train[features]
y_train = df_train[["R1M_Usd_C"]]

X_test = df_test[features]
y_test = df_test[["R1M_Usd_C"]]

#from xgboost import XGBClassifier

#XGBClassifier(

from sklearn.ensemble import RandomForestClassifier

from sklearn.model_selection import GridSearchCV

parameters = {
    #"n_estimators":[100, 200, 300, 400, 500],
    #"n_estimators":[100, 150, 200, 250],
    "n_estimators":[50, 100, 150],
    #"max_depth":[2, 5, 10, 15]
    "max_depth":[3, 4, 5, 6, 7],
    "max_features":["sqrt", None], 
    #"ccp_alpha":[0.0001, 0.001]
}

rf_clf = RandomForestClassifier(n_jobs=-1)

grid_search = GridSearchCV(rf_clf, parameters, verbose=1)

grid_search.fit(X_train, y_train.values.ravel())

df_cv_results = pd.DataFrame(grid_search.cv_results_).sort_values("rank_test_score")
df_cv_results

#df_train

#1 - y_train.mean()

#y_train.mean()

#df_train

df_cv_results.to_csv("../data/random_forest_cv_results.csv", index=False)

pd.read_csv("../data/random_forest_cv_results.csv")

(4 + 5 + 4 + 4 + 6)/5
