==> research/10_xgb_hyperparameter_tuning.ipynb <==
import numpy as np
import pandas as pd

df_ml = pd.read_csv("../data/data_ml.csv")

separation_date = "2013-12-31"
df_train = df_ml.query("date < @separation_date").copy()
df_test = df_ml.query("@separation_date <= date & date < '2018-12-31'").copy()

columns_to_drop = [
    "stock_id", "date", # non-feature identifiers
    "R1M_Usd", "R3M_Usd", "R6M_Usd", "R12M_Usd", # numerical labels
    "R1M_Usd_C", "R12M_Usd_C" # categorical labels
]

features = list(df_ml.drop(columns=columns_to_drop).columns)

X_train = df_train[features]
y_train = df_train["R1M_Usd_C"]

X_test = df_test[features]
y_test = df_test["R1M_Usd_C"]

#from xgboost import XGBClassifier

#XGBClassifier(

#from sklearn.ensemble import RandomForestClassifier
from xgboost import XGBClassifier

from sklearn.model_selection import GridSearchCV

parameters = {
    "n_estimators":[5, 10, 15, 20],
    "colsample_bytree":[0.4, 0.6, 0.8],
    "max_depth": [2, 4, 6, 8, 10],
    "reg_alpha": [0],
    "learning_rate": [0.1, 0.2, 0.3, 0.4, 0.5, 0.6]
}

xgb_clf = XGBClassifier(n_jobs=-1)

grid_search = GridSearchCV(xgb_clf, parameters, refit=True)

grid_search.fit(X_train, y_train)

df_cv_results = pd.DataFrame(grid_search.cv_results_).sort_values("rank_test_score")
df_cv_results

np.mean(grid_search.predict(X_train))

y_train.mean()

grid_search.score(X_train, y_train)

#1 - y_train.mean()

#y_train.mean()

#df_train

#df_cv_results.to_csv("../data/xgboost_cv_results.csv", index=False)

#pd.read_csv("../data/random_forest_cv_results.csv")

## Tuning Notes

`n_estimators=[25, 50, 100, 200]` and got accuracy scores of 0.486 

`n_estimators=[2, 5]` got accuracy scores of 0.506

parameters = {
    "n_estimators":[5],
    "colsample_bytree":[0.3, 0.4, 0.5, 0.6, 0.7, 0.8]
}
accuracy scores around 0.501; 0.6 was the best at 0.508

parameters = {
    "n_estimators":[5],
    "colsample_bytree":[0.6],
    "max_depth": [2, 4, 6, 8, 10, 15, 20]
} - max_depth of 4 had the highest accuracy of 0.51

parameters = {
    "n_estimators":[5],
    "colsample_bytree":[0.6],
    "max_depth": [4],
    "reg_alpha": [0, 0.001, 0.01, 0.1, 0.5],
} - reg_alpha doesn't seem to make a difference, so we'll leave it at zero

Final Tuned Values:
1. colsample_bytree = 0.5
2. learning_rate = 0.4
3. max_depth = 3
4. n_estimators = 10
5. reg_alpha = 0

## References

https://www.analyticsvidhya.com/blog/2016/03/complete-guide-parameter-tuning-xgboost-with-codes-python/

https://medium.com/@jyotsna.a.choudhary/mastering-xgboost-a-technical-guide-for-intermediate-machine-learning-practitioners-f7ad167c6865
