==> research/15_random_forest_untuned_fitting.ipynb <==
## Importing Packages

import numpy as np
import pandas as pd

## Reading and Separating Data

df_ml = pd.read_csv("../data/data_ml.csv")

separation_date = "2013-12-31"
df_train = df_ml.query("date < @separation_date").copy()
df_test = df_ml.query("@separation_date <= date & date < '2018-12-31'").copy()

columns_to_drop = [
    "stock_id", "date", # non-feature identifiers
    "R1M_Usd", "R3M_Usd", "R6M_Usd", "R12M_Usd", # numerical labels
    "R1M_Usd_C", "R12M_Usd_C" # categorical labels
]

features = list(df_ml.drop(columns=columns_to_drop).columns)

X_train = df_train[features]
y_train = df_train["R1M_Usd_C"]

X_test = df_test[features]
y_test = df_test["R1M_Usd_C"]

df_features_selected = pd.read_csv("../data/features_selected.csv")
features_selected = list(df_features_selected["feature"].values)

## Cross-Validation Fitting to All Features

from sklearn.ensemble import RandomForestClassifier
from sklearn.model_selection import cross_val_score

model = RandomForestClassifier(max_depth=5, random_state=0, n_jobs=-1)
cv_scores_all_features = cross_val_score(model, X_train, y_train)
cv_scores_all_features

cv_scores_all_features.mean()

## Cross-Validation Fitting to Selected Features

model = RandomForestClassifier(max_depth=5, random_state=0, n_jobs=-1)
cv_scores_selected_features = cross_val_score(model, X_train[features_selected], y_train)
cv_scores_selected_features

cv_scores_selected_features.mean()

## Final Fitting to All Features

model = RandomForestClassifier(max_depth=5, random_state=0, n_jobs=-1)
model.fit(X_train, y_train)

hard_predictions = model.predict(X_test)
hard_predictions

prob_predict_0 = model.predict_proba(X_test)[:,0]
prob_predict_1 = model.predict_proba(X_test)[:,1]

df_inference_all_features = pd.DataFrame({
    "hard": hard_predictions,
    "probability_0": prob_predict_0,
    "probability_1": prob_predict_1,
})
df_inference_all_features

df_inference_all_features.to_csv("../data/inference_random_forest_untuned_all_features.csv", index=False)

## Final Fitting to Selected Features

model = RandomForestClassifier(max_depth=5, random_state=0, n_jobs=-1)
model.fit(X_train[features_selected], y_train)

hard_predictions = model.predict(X_test[features_selected])
hard_predictions

prob_predict_0 = model.predict_proba(X_test[features_selected])[:,0]
prob_predict_1 = model.predict_proba(X_test[features_selected])[:,1]

df_inference_selected_features = pd.DataFrame({
    "hard": hard_predictions,
    "probability_0": prob_predict_0,
    "probability_1": prob_predict_1,
})
df_inference_selected_features

df_inference_selected_features.to_csv("../data/inference_random_forest_untuned_selected_features.csv", index=False)
