==> research/19_random_forest_tuning_all_features.ipynb <==
import numpy as np
import pandas as pd
import sklearn

df_ml = pd.read_csv("../data/data_ml.csv")

separation_date = "2013-12-31"
df_train = df_ml.query("date < @separation_date").copy()
df_test = df_ml.query("@separation_date <= date & date < '2018-12-31'").copy()

columns_to_drop = [
    "stock_id", "date", # non-feature identifiers
    "R1M_Usd", "R3M_Usd", "R6M_Usd", "R12M_Usd", # numerical labels
    "R1M_Usd_C", "R12M_Usd_C" # categorical labels
]

features = list(df_ml.drop(columns=columns_to_drop).columns)

X_train = df_train[features]
y_train = df_train["R1M_Usd_C"]

from sklearn.ensemble import RandomForestClassifier
from sklearn.model_selection import GridSearchCV

parameters = {
    "n_estimators":[25, 50, 100, 150, 200],
    "max_depth":[5]
}
scoring = [
    "accuracy",
    "precision",
    "recall",
    "roc_auc"
]

model = RandomForestClassifier(n_jobs=-1)

cv = sklearn.model_selection.StratifiedKFold(shuffle=True)
grid_search = GridSearchCV(model, parameters, scoring=scoring, refit=False, verbose=1, cv=cv)

grid_search.fit(X_train, y_train)

df_cv_results = pd.DataFrame(grid_search.cv_results_)#.sort_values("rank_test_score")
#df_cv_results

df_cv_results.info()

cols = [
    "mean_fit_time",         
    # "std_fit_time",          
    # "mean_score_time",       
    # "std_score_time",        
    "param_max_depth",       
    "param_n_estimators",    
    "params",  
    # "split0_test_accuracy",  
    # "split1_test_accuracy",  
    # "split2_test_accuracy",  
    # "split3_test_accuracy",  
    # "split4_test_accuracy",  
    "mean_test_accuracy",    
    # "std_test_accuracy",     
    "rank_test_accuracy",    
    # "split0_test_precision", 
    # "split1_test_precision", 
    # "split2_test_precision", 
    # "split3_test_precision", 
    # "split4_test_precision", 
    "mean_test_precision",   
    # "std_test_precision",    
    "rank_test_precision",   
    # "split0_test_recall",    
    # "split1_test_recall",    
    # "split2_test_recall",    
    # "split3_test_recall",    
    # "split4_test_recall",    
    "mean_test_recall",      
    # "std_test_recall",       
    "rank_test_recall",      
    # "split0_test_roc_auc",   
    # "split1_test_roc_auc",   
    # "split2_test_roc_auc",   
    # "split3_test_roc_auc",   
    # "split4_test_roc_auc",   
    "mean_test_roc_auc",     
    # "std_test_roc_auc",      
    "rank_test_roc_auc",
]

df_cv_results[cols]
