==> research/98_many_models_data_generation.ipynb <==
import numpy as np
import pandas as pd
import sklearn
import itertools

## Reading and Separating Data

df_ml = pd.read_csv("../data/data_ml.csv")

separation_date = "2013-12-31"
df_train = df_ml.query("date < @separation_date").copy()
df_test = df_ml.query("@separation_date <= date & date < '2018-12-31'").copy()

columns_to_drop = [
    "stock_id", "date", # non-feature identifiers
    "R1M_Usd", "R3M_Usd", "R6M_Usd", "R12M_Usd", # numerical labels
    "R1M_Usd_C", "R12M_Usd_C" # categorical labels
]

features = list(df_ml.drop(columns=columns_to_drop).columns)

X_train = df_train[features]
y_train = df_train["R1M_Usd_C"]

X_test = df_test[features]
y_test = df_test["R1M_Usd_C"]

rebalance_dates = list(np.sort(df_test["date"].unique()))

from sklearn.model_selection import cross_validate
from sklearn.linear_model import LogisticRegression

def get_fitting_backtesting_results(specified_model):
    # fitting cross-validation results
    cv = sklearn.model_selection.StratifiedKFold(n_splits=5, shuffle=True)
    scoring = [
        "accuracy",
        "precision",
        "recall",
        "roc_auc"
    ]
    results_cv = cross_validate(specified_model, X_train, y_train, scoring=scoring, cv=cv)
    accuracy_mean = results_cv['test_accuracy'].mean()
    precision_mean = results_cv['test_precision'].mean()
    recall_mean = results_cv['test_recall'].mean()
    roc_auc_mean = results_cv['test_roc_auc'].mean()

    # final fitting
    model = specified_model
    model.fit(X_train, y_train)

    # generating inferences
    df_screen = df_test[["date", "R1M_Usd"]].copy().reset_index(drop=True)
    df_screen["pred_hard"] = model.predict(X_test)
    df_screen["pred_prob"] = model.predict_proba(X_test)[:,1]

    # hard-prediction
    df_monthly_returns = df_screen.query("pred_hard==1").groupby(["date"])[["R1M_Usd"]].mean().reset_index()
    df_monthly_returns['growth_factor'] = 1 + df_monthly_returns['R1M_Usd']
    df_monthly_returns['equity_curve'] = df_monthly_returns['growth_factor'].cumprod()
    annual_return_hp = ((df_monthly_returns['equity_curve'].iloc[-1]) ** (12 / len(df_monthly_returns))) - 1
    sharpe_ratio_hp = (df_monthly_returns["R1M_Usd"].mean() / df_monthly_returns["R1M_Usd"].std()) * np.sqrt(12)
    
    # top-200 trades
    rebalance_dates = list(np.sort(df_screen["date"].unique()))
    lst_trades = []
    for ix_date in rebalance_dates:
        df = df_screen.query("date == @ix_date").sort_values(by="pred_prob", ascending=False).head(200).copy()
        lst_trades.append(df)
    df_trades = pd.concat(lst_trades)
    
    # top-200 return and sharpe-ratio
    df_monthly_returns = df_trades.groupby("date")[["R1M_Usd"]].mean().reset_index()
    df_monthly_returns['growth_factor'] = 1 + df_monthly_returns['R1M_Usd']
    df_monthly_returns['equity_curve'] = df_monthly_returns['growth_factor'].cumprod()
    annual_return_top = ((df_monthly_returns['equity_curve'].iloc[-1]) ** (12 / len(df_monthly_returns))) - 1
    sharpe_ratio_top = (df_monthly_returns["R1M_Usd"].mean() / df_monthly_returns["R1M_Usd"].std()) * np.sqrt(12)

    # putting results into a DataFrame - Random Forest
    df_results = pd.DataFrame({
        "model_family":["random_forest"],
        "model_name":[str(model)],
        "n_estimators":[model.get_params()["n_estimators"]],
        "criterion":[model.get_params()["criterion"]],
        "max_depth":[model.get_params()["max_depth"]],
        "max_features":[model.get_params()["max_features"]],
        "accuracy":[accuracy_mean],
        "precision":[precision_mean],
        "recall":[recall_mean],
        "roc_auc":[roc_auc_mean],
        "hp_ann_ret":[annual_return_hp],
        "hp_sharpe":[sharpe_ratio_hp],
        "top_200_ann_ret":[annual_return_top],
        "top_200_sharpe":[sharpe_ratio_top],
    })

    return df_results

## Logistic Regression

#get_fitting_backtesting_results(LogisticRegression(penalty=None))

# #penalty = ["l2"]
# penalty = [None]
# #C = np.logspace(-5, 1, 7)
# solver = ["lbfgs", "newton-cg", "newton-cholesky", "sag", "saga"]
# results = []
# ix_counter = 1
# for ix_param in itertools.product(penalty, solver):
#     #specified_model = LogisticRegression(penalty=ix_param[0], C=ix_param[1], solver=ix_param[2], max_iter=10000, n_jobs=-1)
#     specified_model = LogisticRegression(penalty=ix_param[0], solver=ix_param[1], max_iter=10000, n_jobs=-1)
#     print(ix_counter, specified_model)
#     df = get_fitting_backtesting_results(specified_model)
#     results.append(df)
#     ix_counter +=  1
# df_results = pd.concat(results)

## Random Forest

from sklearn.ensemble import RandomForestClassifier

df_result = get_fitting_backtesting_results(RandomForestClassifier(max_depth=5, n_jobs=-1))

df_result

n_estimators = [200] #[25, 50, 100, 150, 200]
criterion = ["gini", "entropy", "log_loss"]
max_depth = [2, 4, 6, 8, 10, 12, 14]
max_features = ["sqrt", "log2", None, 0.1, 0.25, 0.5, 0.75]
results = []
ix_counter = 1
for ix_param in itertools.product(n_estimators, criterion, max_depth, max_features):
    specified_model = RandomForestClassifier(
        n_estimators=ix_param[0],
        criterion=ix_param[1],
        max_depth=ix_param[2],
        max_features=ix_param[3],
        n_jobs=-1
    )
    print(ix_counter, specified_model)
    df = get_fitting_backtesting_results(specified_model)
    results.append(df)
    ix_counter +=  1
df_results = pd.concat(results)

# df_results = pd.concat(results)
df_results

# df_results.plot(x="accuracy", y="top_200_ann_ret", kind="scatter");

# df_results.plot(x="precision", y="top_200_ann_ret", kind="scatter");

# df_results.plot(x="recall", y="top_200_ann_ret", kind="scatter");

# df_results.plot(x="roc_auc", y="top_200_ann_ret", kind="scatter");

#np.logspace(-5, 2, 8)
#df_results.to_csv("../data/many_models_random_forest_200.csv", index=False)

# penalty = ["l1", "l2"]
# C = np.logspace(-5, 2, 8)
# solver = ["lbfgs", "newton-cg", "newton-cholesky", "sag"]

# lst = []
# for ix_parameters in itertools.product(penalty, C, solver):
#     lst.append(ix_parameters)
# len(lst)

# df_results["hp_ann_ret"].hist()

# df_results["top_200_ann_ret"].hist(bins=50);

# df_results[["top_200_ann_ret"]].describe()

#LogisticRegression(C=10.0, max_iter=10000, n_jobs=-1, solver='sag')
#LogisticRegression(C=10.0, max_iter=10000, n_jobs=-1, solver='saga')
